consolidates ssh parsing logic
