+++ backend/pkg/sqlconnect/sql-connector_test.go
@@ -142,19 +142,3 @@ func Test_NewDbFromConnectionConfig(t *testing.T) {
 func ptr[T any](val T) *T {
 	return &val
 }
-
-func Test_getSshAddr(t *testing.T) {
-	t.Run("with port", func(t *testing.T) {
-		actual := getSshAddr(&mgmtv1alpha1.SSHTunnel{
-			Host: "localhost",
-			Port: 2222,
-		})
-		assert.Equal(t, "localhost:2222", actual)
-	})
-	t.Run("without port", func(t *testing.T) {
-		actual := getSshAddr(&mgmtv1alpha1.SSHTunnel{
-			Host: "localhost",
-		})
-		assert.Equal(t, "localhost", actual)
-	})
-}
